etl/requirements.txt
psycopg2-binary==2.9.9
elasticsearch==7.17.0
pydantic-settings==2.2.1
backoff==2.2.1